Algo trading/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
apscheduler==3.10.4
pytz==2024.1

# ── Data Feeds ────────────────────────────────────────────────────────────────
alpaca-py==0.20.0
ccxt==4.2.25
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
websocket-client==1.7.0

# ── Technical Analysis ────────────────────────────────────────────────────────
ta==0.11.0
pandas-ta==0.3.14b
scipy==1.12.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
tensorflow==2.15.0
keras==2.15.0
joblib==1.3.2

# ── Quantitative Finance ──────────────────────────────────────────────────────
quantlib==1.33
PyPortfolioOpt==1.5.5
statsmodels==0.14.1

# ── Reporting ─────────────────────────────────────────────────────────────────
openpyxl==3.1.2
xlsxwriter==3.1.9

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.31.0
plotly==5.18.0
altair==5.2.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
aiosqlite==0.19.0

# ── Utilities ─────────────────────────────────────────────────────────────────
loguru==0.7.2
pydantic==2.6.0
python-dateutil==2.8.2
tqdm==4.66.1
colorama==0.4.6